# backend_go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/categories/slug/{slug}": {
            "get": {
                "description": "Retrieve a specific job category by slug",
                "tags": [
                    "categories"
                ],
                "summary": "Get job category by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{category_id}/trades": {
            "get": {
                "description": "Retrieve job trades for a specific category",
                "tags": [
                    "trades"
                ],
                "summary": "Get job trades by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{category_id}/trades/{slug}": {
            "get": {
                "description": "Retrieve a specific job trade by slug and category",
                "tags": [
                    "trades"
                ],
                "summary": "Get job trade by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job trade slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "description": "Retrieve a specific job category by ID",
                "tags": [
                    "categories"
                ],
                "summary": "Get job category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications/{id}": {
            "get": {
                "description": "Retrieve a specific notification log by ID",
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}/clicked": {
            "post": {
                "description": "Mark a notification as clicked",
                "tags": [
                    "notifications"
                ],
                "summary": "Mark notification as clicked",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}/read": {
            "post": {
                "description": "Mark a notification as read",
                "tags": [
                    "notifications"
                ],
                "summary": "Mark notification as read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/trades": {
            "get": {
                "description": "Retrieve paginated list of job trades with optional filters",
                "tags": [
                    "trades"
                ],
                "summary": "Get job trades",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by demand level",
                        "name": "demand_level",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/trades/{id}": {
            "get": {
                "description": "Retrieve a specific job trade by ID",
                "tags": [
                    "trades"
                ],
                "summary": "Get job trade by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job trade ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{user_id}/notifications": {
            "get": {
                "description": "Retrieve paginated list of user notification logs with optional filters",
                "tags": [
                    "notifications"
                ],
                "summary": "Get user notification logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by notification type",
                        "name": "notification_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notification log",
                "tags": [
                    "notifications"
                ],
                "summary": "Create notification log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateNotificationRequest": {
            "type": "object",
            "required": [
                "channel",
                "message",
                "notification_type",
                "title"
            ],
            "properties": {
                "channel": {
                    "type": "string"
                },
                "enrollment_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "notification_type": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": true
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}